% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosx.R
\name{print.cosx}
\alias{print.cosx}
\title{Print R cosx fortunes.}
\usage{
\method{print}{cosx}(x, width = NULL, ...)
}
\arguments{
\item{x}{an object of class "cosx", usually a single row from `cosxs.data`.}

\item{width}{integer specifying the character width. By default getOption("width") is used.}

\item{...}{potential further arguments passed to `grep`.}
}
\value{
print.
}
\description{
Print R cosx fortunes.
}
